@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "digiarch"
-version = "1.5.0"
+version = "2.0.0"
 description = "Tools for the Digital Archive Project at Aarhus Stadsarkiv"
 authors = ["Aryan Muhammadi Landi <arml@aarhus.dk>", "Nina Jensen <jnik@aarhus.dk>", "Aarhus Stadsarkiv <stadsarkiv@aarhus.dk>"]
 license = "GPL-3.0"
@@ -11,26 +11,26 @@ keywords = ["archiving", "preservation", "identification"]
 include = ["pyproject.toml"]
 
 [tool.poetry.dependencies]
-python = "^3.9"
-acacore = {git = "https://github.com/aarhusstadsarkiv/acacore.git", tag = "v1.2.0"}
+python = "^3.11"
+acacore = {git = "https://github.com/aarhusstadsarkiv/acacore.git", tag = "v2.0.0"}
 
 [tool.poetry.group.dev.dependencies]
-pytest = "^8.2.1"
-black = "^24.4.2"
-ruff = "^0.4.6"
-coverage = "^7.5.3"
+pytest = "^8.3.2"
+ruff = "^0.5.6"
+coverage = "^7.6.1"
 types-click = "^7.1.8"
 
 [tool.poetry.scripts]
 digiarch = "digiarch.cli:app"
 
-[tool.black]
-line-length = 120
-exclude = ".venv"
-
 [tool.ruff]
 line-length = 120
-lint.select = [
+
+[tool.ruff.format]
+line-ending = "lf"
+
+[tool.ruff.lint]
+select = [
     "C",
     "E",
     "F",
@@ -47,7 +47,7 @@ lint.select = [
     "BLE",
     # "FBT", Will asks us to remove boolean flags
     "A",
-    "COM",
+    # "COM",
     "C4",
     "DTZ",
     "T10",
@@ -67,7 +67,7 @@ lint.select = [
     # "TRY", Will help with our exception handling
     "RUF",
 ]
-lint.ignore = [
+ignore = [
     "E501",    # line to long, to many false positives, gets handled by black
     "ANN002",  # missing type anottation for *args
     "ANN003",  # missing type anotation for **kwargs
@@ -91,6 +91,7 @@ lint.ignore = [
     "D212",    # multi-line docstring summary should start at the first line,
     "D213",    # multi-line docstring summary should start at the second line,
     "D300",    # use triple docstring
+    "D401",    # first line of docstring should be in imperative mood
     "DTZ005",  # datetime.datetime.now() called without a tz argument
     "DTZ006",  # datetime.datetime.fromtimestamp() called without a tz argument
     "E712",    # comparison to True/False, we ignore because we use sqlalchemy
@@ -125,27 +126,11 @@ force-single-line = true
 order-by-type = false
 
 [tool.ruff.lint.pydocstyle]
-convention = "google"
+convention = "pep257"
 ignore-decorators = ["pytest.fixture"]
 
 [tool.ruff.lint.pycodestyle]
-max-doc-length = 300 # We dont care about length of docstrings
-
-[tool.mypy]
-python_version = 3.9
-warn_return_any = true
-warn_unused_configs = true
-ignore_missing_imports = true
-# Because we disallow untyped defs in the main module,
-# check_untyped_defs only has an impact in tests/
-check_untyped_defs = true
-exclude = [
-    ".venv/"
-]
-
-[[tool.mypy.overrides]]
-module = 'digiarch'
-disallow_untyped_defs = true
+max-doc-length = 300
 
 [build-system]
 requires = ["poetry>=0.12"]
